+++ src/Smartstore.Core/Content/Media/Bootstrapping/MediaStarter.cs
@@ -54,6 +54,11 @@ namespace Smartstore.Core.Bootstrapping
             builder.RegisterType<DownloadService>().As<IDownloadService>().InstancePerLifetimeScope();
             builder.RegisterType<ImageOffloader>().As<IImageOffloder>().InstancePerLifetimeScope();
 
+            // Duplicate detectors.
+            builder.RegisterType<DefaultMediaDupeDetector>().As<IMediaDupeDetector>().InstancePerLifetimeScope();
+            builder.RegisterType<CachingMediaDupeDetector>().As<IMediaDupeDetector>().InstancePerLifetimeScope();
+            builder.RegisterType<MediaDupeDetectorFactory>().As<IMediaDupeDetectorFactory>().InstancePerLifetimeScope();
+
             // ImageSharp adapter factory
             builder.RegisterType<SharpImageFactory>().As<IImageFactory>().SingleInstance();
             builder.RegisterType<ImageCache>().As<IImageCache>().InstancePerLifetimeScope();